requirements.txt
PyPDF2==3.0.1 
pdfplumber==0.10.3 
requests==2.31.0 
pyyaml==6.0.1 
google-api-python-client==2.95.0 
google-auth-httplib2==0.1.0 
google-auth-oauthlib==1.0.0 
jaconv==0.3.4 
python-dotenv==1.0.0 
colorlog==6.7.0 
